Correct how flash messages are set and read
Initially, I misunderstood how the gorilla/sesisons package interacted
with the session, leading to adding flash messages which were never
unset on read; which the package states, in places, is what happens.

This change corrects that problem so that when a flash message is read,
the underlysing storage, the session cookie, is updated with the new
flash messages list; the original list, minus the one just read.

In addition, this change:

- Stores flash messages under an application-specific key, so that there
  is no namespace collision with any other user of the session
- The session isn't polluted or hard to discern as everything is dumped
  in to the one session namespace
- Where the application's flash messages are stored is easier to change,
  as it's based on a single configuration value